+++ pyccel/ast/cwrapper.py
@@ -9,6 +9,7 @@ between Python code and C code (using Python/C Api and cwrapper.c).
 This file contains classes but also many FunctionDef/Variable instances representing
 objects defined in Python.h.
 """
+import re
 
 from pyccel.utilities.metaclasses import Singleton
 
@@ -55,6 +56,7 @@ __all__ = (
     'PyModule',
     'PyArgKeywords',
     'PyArg_ParseTupleNode',
+    'PyArgumentError',
     'PyBuildValueNode',
     'PyCapsule_New',
     'PyCapsule_Import',
@@ -1028,6 +1030,71 @@ class PyTuple_Pack(PyccelFunction):
     __slots__ = ()
     _class_type = PyccelPyObject()
     _shape = None
+
+class PyArgumentError(PyccelAstNode):
+    """
+    Class to display errors related to arguments.
+
+    Class to display errors related to arguments. This class helps
+    format the arguments to display the type of the received argument.
+
+    Parameters
+    ----------
+    error_type : Variable
+        A Variable containing the error type to be raised. E.g. PyTypeError.
+    error_msg : str
+        The message to be displayed containing f-string style type indicators.
+    **kwargs : dict[str, Variable]
+        The arguments whose types will be printed.
+    """
+    __slots__ = ('_error_type', '_error_msg', '_args')
+    _attribute_nodes = ('_args',)
+
+    def __init__(self, error_type, error_msg : str, **kwargs):
+        assert isinstance(error_type, Variable)
+        assert isinstance(error_msg, str)
+        args = []
+        # Find all expressions of the style '{type(var_name)}' in the error message
+        type_indicators = re.findall(r'{type\([a-zA-Z0-9_]+\)}', error_msg)
+        # Save the error message, replacing type indicators with the format string
+        self._error_msg = re.sub(r'{type\([a-zA-Z0-9_]+\)}', '%V', error_msg)
+        # Find the relevant arguments for each type indicator
+        for t in type_indicators:
+            var_name = t.removeprefix('{type(').removesuffix(')}')
+            args.append(ObjectAddress(kwargs[var_name]))
+
+        self._args = tuple(args)
+        self._error_type = error_type
+        super().__init__()
+
+    @property
+    def error_type(self):
+        """
+        The error type that should be raised.
+
+        The error type that should be raised.
+        """
+        return self._error_type
+
+    @property
+    def error_msg(self):
+        """
+        The error message that should be formatted.
+
+        The error message that should be formatted.
+        """
+        return self._error_msg
+
+    @property
+    def args(self):
+        """
+        The arguments whose types are printed in the error message.
+
+        The arguments whose types are printed in the error message.
+        These arguments are displayed in the order they appear in
+        the error message.
+        """
+        return self._args
 
 #-------------------------------------------------------------------
 #                      Python.h Constants