+++ pyccel/parser/semantic.py
@@ -355,6 +355,46 @@ class SemanticParser(BasicParser):
     #              Utility functions for scope handling
     #================================================================
 
+    def create_new_function_scope(self, syntactic_name, semantic_name, **kwargs):
+        """
+        Create a new Scope object for a Python function.
+
+        Create a new Scope object for a Python function with the given name,
+        and attach any decorators' information to the scope. The new scope is
+        a child of the current one, and can be accessed from the dictionary of
+        its children using the function name as key.
+
+        Before returning control to the caller, the current scope (stored in
+        self._scope) is changed to the one just created, and the function's
+        name is stored in self._current_function_name.
+
+        Parameters
+        ----------
+        syntactic_name : str
+            Function's original name in the translated code, used as a key to
+            retrieve the new scope.
+
+        semantic_name : str
+            The new name of the function by which it will be known in the target
+            language.
+
+        **kwargs : dict
+            Keyword arguments passed through to the new scope.
+
+        Returns
+        -------
+        Scope
+            The new scope for the function.
+        """
+        child = self.scope.new_child_scope(syntactic_name, **kwargs)
+        child.local_used_symbols[syntactic_name] = semantic_name
+        child.python_names[semantic_name] = syntactic_name
+
+        self._scope = child
+        self._current_function_name.append(semantic_name)
+
+        return child
+
     def get_class_prefix(self, name):
         """
         Search for the class prefix of a dotted name in the current scope.
@@ -629,7 +669,7 @@ class SemanticParser(BasicParser):
         if all(r.expr is None for r in expr.get_attribute_nodes(Return)):
             for i in self._allocs[-1]:
                 if isinstance(i, DottedVariable):
-                    if isinstance(i.lhs.class_type, CustomDataType) and self._current_function != '__del__':
+                    if isinstance(i.lhs.class_type, CustomDataType) and self.current_function_name != '__del__':
                         continue
                 if isinstance(i.class_type, CustomDataType) and i.is_alias:
                     continue
@@ -663,28 +703,22 @@ class SemanticParser(BasicParser):
             A list of objects in `_allocs` which are to be ignored (variables appearing
             in a return statement).
         """
+        assert len(self._allocs) == len(self._pointer_targets)
         assert not isinstance(exceptions, Variable)
         for i in self._allocs[-1]:
-            if isinstance(i, DottedVariable):
-                if isinstance(i.lhs.class_type, CustomDataType) and self._current_function != '__del__':
-                    continue
             if i in exceptions:
                 continue
             self._pointer_targets[-1].pop(i, None)
         targets = {t[0]:t[1] for target_list in self._pointer_targets[-1].values() for t in target_list}
         for i in self._allocs[-1]:
-            if isinstance(i, DottedVariable):
-                if isinstance(i.lhs.class_type, CustomDataType) and self._current_function != '__del__':
-                    continue
             if i in exceptions:
                 continue
             if i in targets:
                 errors.report(f"Variable {i} goes out of scope but may be the target of a pointer which is still required",
                         severity='error', symbol=targets[i])
 
-        if self._current_function:
-            func_name = self._current_function.name[-1] if isinstance(self._current_function, DottedName) else self._current_function
-            current_func = self.scope.find(func_name, 'functions')
+        if self.current_function_name:
+            current_func = self._current_function[-1]
             arg_vars = {a.var:a for a in current_func.arguments}
 
             for p, t_list in self._pointer_targets[-1].items():
@@ -1106,6 +1140,72 @@ class SemanticParser(BasicParser):
 
             return PythonTuple(*([self.scope.collect_tuple_element(v) for v in val]*length))
 
+    def _create_class_destructor(self, expr):
+        """
+        Create the class destructor.
+
+        Create the class destructor. This is important to ensure that the data in the
+        class is correctly deallocated. If the class already has a destructor then the
+        deallocations are added to the existing destructor. Similarly a flag is added
+        to the init method to act as a guard in the class to tell if the destructor has
+        been called.
+
+        Parameters
+        ----------
+        expr : ClassDef
+            The class that implicit __init__ and __del__ methods should be created for.
+        """
+        class_type = expr.class_type
+        methods = expr.methods
+        cls_scope = expr.scope
+
+        init_func = cls_scope.functions['__init__']
+
+        if isinstance(init_func, Interface):
+            errors.report("Pyccel does not support interface constructor", symbol=init_func,
+                severity='fatal')
+
+        # create a new attribute to check allocation
+        deallocater_lhs = Variable(class_type, 'self', cls_base = expr, is_argument=True)
+        deallocater = DottedVariable(lhs = deallocater_lhs, name = self.scope.get_new_name('is_freed'),
+                                     class_type = PythonNativeBool(), is_private=True)
+        expr.add_new_attribute(deallocater)
+        deallocater_assign = Assign(deallocater, LiteralFalse())
+        init_func.body.insert2body(deallocater_assign, back=False)
+
+        del_method = next((method for method in methods if method.name == '__del__'), None)
+        if del_method is None:
+            argument = FunctionDefArgument(Variable(class_type, 'self', cls_base = expr), bound_argument = True)
+            del_name = cls_scope.get_new_name('__del__')
+            scope = self.create_new_function_scope('__del__', del_name)
+            scope.insert_variable(argument.var)
+            self.exit_function_scope()
+            del_method = FunctionDef(del_name, [argument], [Pass()], scope=scope)
+            self.insert_function(del_method, cls_scope)
+            expr.add_new_method(del_method)
+        else:
+            assert del_method.is_semantic
+
+        # Add destructors to __del__ method
+        self._current_function_name.append(del_method.name)
+        attribute = []
+        for attr in expr.attributes:
+            if not attr.on_stack:
+                attribute.append(attr)
+            elif isinstance(attr.class_type, CustomDataType) and not attr.is_alias:
+                attribute.append(attr)
+        if attribute:
+            # Create a new list that store local attributes
+            self._allocs.append(set())
+            self._pointer_targets.append({})
+            self._allocs[-1].update(attribute)
+            del_method.body.insert2body(*self._garbage_collector(del_method.body))
+            self._pointer_targets.pop()
+        condition = If(IfSection(PyccelNot(deallocater),
+                        [del_method.body]+[Assign(deallocater, LiteralTrue())]))
+        del_method.body = [condition]
+        self._current_function_name.pop()
+
     def _handle_function_args(self, arguments):
         """
         Get a list of all function arguments.
@@ -1131,7 +1231,7 @@ class SemanticParser(BasicParser):
             a = self._visit(arg)
             val = a.value
             if isinstance(val, FunctionDef) and not isinstance(val, PyccelFunctionDef) and not val.is_semantic:
-                semantic_func = self._annotate_the_called_function_def(val)
+                semantic_func = self._annotate_the_called_function_def(val, ())
                 a = FunctionCallArgument(semantic_func, keyword = a.keyword, python_ast = a.python_ast)
 
             if isinstance(val, StarredArguments):
@@ -1270,7 +1370,18 @@ class SemanticParser(BasicParser):
 
             return new_expr
         else:
-            if self._current_function == func.name:
+            is_inline = func.is_inline if isinstance(func, FunctionDef) else func.functions[0].is_inline
+            if not func.is_semantic:
+                func = self._annotate_the_called_function_def(func, args)
+            elif is_inline and isinstance(func, Interface):
+                is_compatible = False
+                for f in func.functions:
+                    fl = self._check_argument_compatibility(args, f.arguments, func, f.is_elemental, raise_error=False)
+                    is_compatible |= fl
+                if not is_compatible:
+                    func = self._annotate_the_called_function_def(func, args)
+
+            if self.current_function_name == func.name:
                 if func.results and not isinstance(func.results.var, TypedAstNode):
                     errors.report(RECURSIVE_RESULTS_REQUIRED, symbol=func, severity="fatal")
 
@@ -1292,7 +1403,7 @@ class SemanticParser(BasicParser):
                         symbol = expr,
                         severity='fatal')
 
-            new_expr = FunctionCall(func, args, self._current_function)
+            new_expr = FunctionCall(func, args, self.current_function_name)
             for a, f_a in zip(new_expr.args, func_args):
                 if f_a.persistent_target:
                     assert is_method
@@ -1347,7 +1458,7 @@ class SemanticParser(BasicParser):
 
         return input_args
 
-    def _annotate_the_called_function_def(self, old_func, function_call_args=None):
+    def _annotate_the_called_function_def(self, old_func, function_call_args):
         """
         Annotate the called FunctionDef.
 
@@ -1358,7 +1469,7 @@ class SemanticParser(BasicParser):
         old_func : FunctionDef|Interface
            The function that needs to be annotated.
 
-        function_call_args : list[FunctionCallArgument], optional
+        function_call_args : list[FunctionCallArgument]
            The list of the call arguments.
 
         Returns
@@ -1366,32 +1477,47 @@ class SemanticParser(BasicParser):
         func: FunctionDef|Interface
             The new annotated function.
         """
+        cls_base_syntactic = old_func.get_direct_user_nodes(lambda p: isinstance(p, ClassDef))
+        if cls_base_syntactic:
+            cls_name = cls_base_syntactic[0].name
+            cls_base = self.scope.find(cls_name, 'classes')
+            cls_scope = cls_base.scope
+            new_scope = cls_scope
+        else:
+            func_scope = old_func.scope if isinstance(old_func, FunctionDef) else old_func.syntactic_node.scope
+            new_scope = func_scope
         # The function call might be in a completely different scope from the FunctionDef
         # Store the current scope and go to the parent scope of the FunctionDef
-        old_scope            = self._scope
+        old_scope = self._scope
         old_current_function = self._current_function
-        names = []
-        sc = old_func.scope if isinstance(old_func, FunctionDef) else old_func.syntactic_node.scope
-        while sc.parent_scope is not None:
-            sc = sc.parent_scope
-            if not sc.name is None:
-                names.append(sc.name)
-        names.reverse()
-        if names:
-            self._current_function = DottedName(*names) if len(names)>1 else names[0]
-        else:
-            self._current_function = None
+        old_current_function_name = self._current_function_name
 
-        while names:
-            sc = sc.sons_scopes[names[0]]
-            names = names[1:]
+        # Walk up scope to root to find names of relevant scopes
+        scope_names = []
+        while new_scope.parent_scope is not None:
+            new_scope = new_scope.parent_scope
+            if not new_scope.name is None:
+                scope_names.append(new_scope.name)
+
+        # Use scope_names to find semantic scopes
+        for n in scope_names[::-1]:
+            new_scope = new_scope.sons_scopes[n]
 
         # Set the Scope to the FunctionDef's parent Scope and annotate the old_func
-        self._scope = sc
-        self._visit_FunctionDef(old_func, function_call_args=function_call_args)
-        new_name = self.scope.get_expected_name(old_func.name)
+        self._scope = new_scope
+        if old_func.is_inline:
+            self._visit_FunctionDef(old_func, function_call_args = function_call_args)
+        else:
+            self._visit_FunctionDef(old_func)
+
         # Retrieve the annotated function
-        func = self.scope.find(new_name, 'functions')
+        if cls_base_syntactic:
+            new_name = cls_scope.get_expected_name(old_func.name)
+            func = cls_scope.find(new_name, 'functions')
+        else:
+            new_name = self.scope.get_expected_name(old_func.name)
+            func = self.scope.find(new_name, 'functions')
+        assert func is not None
         # Add the Module of the imported function to the new function
         if old_func.is_imported:
             mod = old_func.get_direct_user_nodes(lambda x: isinstance(x, Module))[0]
@@ -1399,6 +1525,7 @@ class SemanticParser(BasicParser):
 
         # Go back to the original Scope
         self._scope = old_scope
+        self._current_function_name = old_current_function_name
         self._current_function = old_current_function
         # Remove the old_func from the imports dict and Assign the new annotated one
         if old_func.is_imported:
@@ -1409,6 +1536,7 @@ class SemanticParser(BasicParser):
             func = func.clone(new_name, is_imported=True)
             func.set_current_user_node(mod)
             scope.imports['functions'][new_name] = func
+
         return func
 
     def _create_variable(self, name, class_type, rhs, d_lhs, *, arr_in_multirets=False,
@@ -1619,7 +1747,8 @@ class SemanticParser(BasicParser):
                 prefix = self.get_class_prefix(lhs)
                 class_def = prefix.cls_base
                 attr_name = lhs.name[-1]
-                attribute = class_def.scope.find(attr_name) if class_def else None
+                attribute = class_def.scope.variables.get(attr_name, None) \
+                            if class_def else None
                 if attribute:
                     var = attribute.clone(attribute.name, new_class = DottedVariable, lhs = prefix)
                 else:
@@ -2586,13 +2715,23 @@ class SemanticParser(BasicParser):
             # it will add the necessary Deallocate nodes
             # to the ast
             program_body.insert2body(*self._garbage_collector(program_body))
+            self._pointer_targets.pop()
 
             self.scope = mod_scope
 
-        for f in self.scope.functions.copy().values():
+        funcs_to_visit = list(self.scope.functions.values())
+        funcs_to_visit.extend(m for c in self.scope.classes.values() for m in c.methods)
+
+        for f in funcs_to_visit:
             if not f.is_semantic and not isinstance(f, InlineFunctionDef):
                 assert isinstance(f, FunctionDef)
                 self._visit(f)
+
+        for c in self.scope.classes.values():
+            self._create_class_destructor(c)
+
+        for f in self.scope.functions.values():
+            assert f.is_semantic or f.is_inline
 
         variables = self.get_variables(self.scope)
         init_func = None
@@ -2605,9 +2744,10 @@ class SemanticParser(BasicParser):
             # If there are any initialisation statements then create an initialisation function
             init_var = Variable(PythonNativeBool(), self.scope.get_new_name('initialised'),
                                 is_private=True, is_temp = True)
-            init_func_name = self.scope.get_new_name(name_suffix+'__init')
+            syntactic_init_func_name = name_suffix+'__init'
+            init_func_name = self.scope.get_new_name(syntactic_init_func_name)
             # Ensure that the function is correctly defined within the namespaces
-            init_scope = self.create_new_function_scope(init_func_name)
+            init_scope = self.create_new_function_scope(syntactic_init_func_name, init_func_name)
             for b in init_func_body:
                 if isinstance(b, ScopedAstNode):
                     b.scope.update_parent_scope(init_scope, is_loop = True)
@@ -2646,7 +2786,8 @@ class SemanticParser(BasicParser):
             self.insert_function(init_func)
 
         if init_func:
-            free_func_name = self.scope.get_new_name(name_suffix+'__free')
+            syntactic_free_func_name = name_suffix+'__free'
+            free_func_name = self.scope.get_new_name(syntactic_free_func_name)
             pyccelised_imports = [imp for imp_name, imp in self.scope.imports['imports'].items() \
                              if imp_name in self.d_parsers]
 
@@ -2664,7 +2805,7 @@ class SemanticParser(BasicParser):
                 free_func_body = If(IfSection(init_var,
                     import_free_calls+deallocs+[Assign(init_var, LiteralFalse())]))
                 # Ensure that the function is correctly defined within the namespaces
-                scope = self.create_new_function_scope(free_func_name)
+                scope = self.create_new_function_scope(syntactic_free_func_name, free_func_name)
                 free_func = FunctionDef(free_func_name, [], [free_func_body],
                                     global_vars = variables, scope = scope)
                 self.exit_function_scope()
@@ -2691,8 +2832,7 @@ class SemanticParser(BasicParser):
                     if F is None:
                         func_defs = []
                         for v in headers:
-                            scope = self.create_new_function_scope(name)
-                            scope.insert_symbol(name)
+                            scope = self.create_new_function_scope(name, v.name)
                             types = [self._visit(d).type_list[0] for d in v.dtypes]
                             args = [Variable(t.class_type, PyccelSymbol(f'anon_{i}'),
                                 shape = None, is_const = t.is_const, is_optional = False,
@@ -2860,6 +3000,12 @@ class SemanticParser(BasicParser):
                     # This leads to printing errors
                     errors.report("Inhomogeneous tuples are not yet supported as arguments",
                             severity='error', symbol=expr)
+                if isinstance(dtype, CustomDataType) and not bound_argument:
+                    cls = self.scope.find(str(dtype), 'classes')
+                    if cls:
+                        init_method = cls.get_method('__init__', expr)
+                        if not init_method.is_semantic:
+                            self._visit(init_method)
                 clone_var = v.clone(v.name, is_optional = is_optional, is_argument = True)
                 args.append(FunctionDefArgument(clone_var, bound_argument = bound_argument,
                                         value = value, kwonly = kwonly, annotation = expr.annotation))
@@ -3122,7 +3268,7 @@ class SemanticParser(BasicParser):
         for t in types.type_list:
             if isinstance(t, FunctionTypeAnnotation):
                 args = t.args
-                scope = self.create_new_function_scope(name)
+                scope = self.create_new_function_scope(name, name)
                 if t.result.var:
                     results = FunctionDefResult(t.result.var.clone(t.result.var.name, is_argument = False),
                                     annotation=t.result.annotation)
@@ -3168,7 +3314,7 @@ class SemanticParser(BasicParser):
         if var is not None:
             new_var = possible_args[0]
             if len(possible_args) != 1 or new_var.class_type != var.class_type:
-                errors.report(f"Variable was declared as the result of the function {self.current_function} but is now declared with a different type",
+                errors.report(f"Variable was declared as the result of the function {self.current_function_name} but is now declared with a different type",
                         symbol=expr, severity='error')
             # Remove variable from scope as AnnotatedPyccelSymbol is always inserted into scope
             self.scope.remove_variable(var)
@@ -3290,20 +3436,15 @@ class SemanticParser(BasicParser):
                 args = [lhs] + list(args)
                 args = [self._visit(i) for i in args]
                 args = macro.apply(args)
-                return FunctionCall(master, args, self._current_function)
+                return FunctionCall(master, args, self.current_function_name)
 
             method = cls_base.get_method(rhs_name, expr)
 
             args = [FunctionCallArgument(visited_lhs), *self._handle_function_args(rhs.args)]
-            if not method.is_semantic:
-                if not method.is_inline:
-                    method = self._annotate_the_called_function_def(method)
-                else:
-                    method = self._annotate_the_called_function_def(method, function_call_args=args)
-
             if cls_base.name == 'numpy.ndarray':
                 numpy_class = method.cls_name
                 self.insert_import('numpy', AsName(numpy_class, numpy_class.name))
+
             return self._handle_function(expr, method, args, is_method = True)
 
         # look for a class attribute / property
@@ -3316,12 +3457,6 @@ class SemanticParser(BasicParser):
             # class property?
             else:
                 method = cls_base.get_method(rhs_name, expr)
-                if not method.is_semantic:
-                    if not method.is_inline:
-                        method = self._annotate_the_called_function_def(method)
-                    else:
-                        method = self._annotate_the_called_function_def(method,
-                                    function_call_args=(FunctionCallArgument(visited_lhs),))
                 assert 'property' in method.decorators
                 if cls_base.name == 'numpy.ndarray':
                     numpy_class = method.cls_name
@@ -3338,7 +3473,7 @@ class SemanticParser(BasicParser):
                 return macro.master
             elif isinstance(macro, MacroFunction):
                 args = macro.apply([visited_lhs])
-                return FunctionCall(macro.master, args, self._current_function)
+                return FunctionCall(macro.master, args, self.current_function_name)
 
         # did something go wrong?
         return errors.report(f'Attribute {rhs_name} not found',
@@ -3429,7 +3564,7 @@ class SemanticParser(BasicParser):
         contains_method = container_base.get_method('__contains__',
                         raise_error_from = expr if isinstance(container_type, CustomDataType) else None)
         if contains_method:
-            return contains_method(container, element)
+            return self._handle_function(expr, contains_method, [FunctionCallArgument(container), FunctionCallArgument(element)])
         else:
             raise errors.report(f"In operator is not yet implemented for type {container_type}",
                     severity='fatal', symbol=expr)
@@ -3490,19 +3625,6 @@ class SemanticParser(BasicParser):
                             annotation=a.annotation, value=a.value, kwonly=a.is_kwonly, bound_argument=a.bound_argument)
                             for a in func_args]
             args      = self._sort_function_call_args(func_args, args)
-            is_inline = func.is_inline if isinstance(func, FunctionDef) else func.functions[0].is_inline
-            if not func.is_semantic:
-                if not is_inline:
-                    func = self._annotate_the_called_function_def(func)
-                else:
-                    func = self._annotate_the_called_function_def(func, function_call_args=args)
-            elif is_inline and isinstance(func, Interface):
-                is_compatible = False
-                for f in func.functions:
-                    fl = self._check_argument_compatibility(args, f.arguments, func, f.is_elemental, raise_error=False)
-                    is_compatible |= fl
-                if not is_compatible:
-                    func = self._annotate_the_called_function_def(func, function_call_args=args)
 
         if name == 'lambdify':
             args = self.scope.find(str(expr.args[0]), 'symbolic_functions')
@@ -3516,6 +3638,9 @@ class SemanticParser(BasicParser):
             d_methods = cls.methods_as_dict
             method = d_methods.pop('__init__', None)
 
+            if not method.is_semantic:
+                method = self._annotate_the_called_function_def(method, args)
+
             if method is None:
 
                 # TODO improve case of class with the no __init__
@@ -3523,8 +3648,8 @@ class SemanticParser(BasicParser):
                 errors.report(UNDEFINED_INIT_METHOD, symbol=name,
                     bounding_box=(self.current_ast_node.lineno, self.current_ast_node.col_offset),
                     severity='error')
-            dtype = method.arguments[0].var.class_type
-            cls_def = method.arguments[0].var.cls_base
+            dtype = cls.class_type
+            cls_def = cls
             d_var = {'class_type' : dtype,
                     'memory_handling':'stack',
                     'shape' : None,
@@ -3551,6 +3676,7 @@ class SemanticParser(BasicParser):
                                     is_augassign = False)
             self._additional_exprs[-1].extend(new_expression)
             args = (FunctionCallArgument(cls_variable), *args)
+
             self._check_argument_compatibility(args, method.arguments,
                             method, method.is_elemental)
 
@@ -3630,10 +3756,11 @@ class SemanticParser(BasicParser):
                 cls_def = semantic_lhs_var.lhs.cls_base
                 insert_scope = cls_def.scope
                 cls_def.add_new_attribute(semantic_lhs_var)
-                lhs = lhs.name.name[-1]
+                lhs_scope_name = lhs.name.name[-1]
             else:
                 insert_scope = self.scope
-                lhs = lhs.name
+                lhs_scope_name = lhs.name
+            lhs = lhs.name
 
             if semantic_lhs_var.class_type is TypeAlias():
                 pyccel_stage.set_stage('syntactic')
@@ -3654,7 +3781,7 @@ class SemanticParser(BasicParser):
                 return EmptyNode()
 
             try:
-                insert_scope.insert_variable(semantic_lhs_var, lhs)
+                insert_scope.insert_variable(semantic_lhs_var, lhs_scope_name)
             except RuntimeError as e:
                 errors.report(e, symbol=expr, severity='error')
 
@@ -3803,7 +3930,7 @@ class SemanticParser(BasicParser):
                 new_expressions += expr
                 args = macro.apply(args, results=results)
                 if isinstance(master.funcdef, FunctionDef):
-                    func_call = FunctionCall(master.funcdef, args, self._current_function)
+                    func_call = FunctionCall(master.funcdef, args, self.current_function_name)
                     if new_expressions:
                         return CodeBlock([*new_expressions, func_call])
                     else:
@@ -4614,11 +4741,11 @@ class SemanticParser(BasicParser):
     def _visit_Return(self, expr):
 
         results     = expr.expr
-        f_name      = self._current_function
+        f_name      = self.current_function_name
         if isinstance(f_name, DottedName):
             f_name = f_name.name[-1]
 
-        func = self.scope.find(f_name, 'functions')
+        func = self._current_function[-1]
 
         original_name = self.scope.get_python_name(f_name)
         if original_name.startswith('__i') and ('__'+original_name[3:]) in magic_method_map.values():
@@ -4687,14 +4814,27 @@ class SemanticParser(BasicParser):
             errors.report("Functions can only be declared in modules or inside other functions.",
                     symbol=expr, severity='error')
 
+        current_class = expr.get_direct_user_nodes(lambda u: isinstance(u, ClassDef))
+        cls_name = current_class[0].name if current_class else None
+        insertion_scope = self.scope
+        if cls_name:
+            bound_class = self.scope.find(cls_name, 'classes', raise_if_missing = True)
+            insertion_scope = bound_class.scope
+
         existing_semantic_funcs = []
         if not expr.is_semantic:
-            self.scope.functions.pop(self.scope.get_expected_name(expr.name), None)
+            name = expr.scope.get_expected_name(expr.name)
+            func = insertion_scope.functions.get(name, None)
+            if func:
+                if func.is_semantic:
+                    return EmptyNode()
+                else:
+                    insertion_scope.functions.pop(name)
         elif isinstance(expr, Interface):
             existing_semantic_funcs = [*expr.functions]
-            expr                    = expr.syntactic_node
+            expr = expr.syntactic_node
+            name = expr.scope.get_expected_name(expr.name)
 
-        name               = self.scope.get_expected_name(expr.name)
         decorators         = expr.decorators
         new_semantic_funcs = []
         sub_funcs          = []
@@ -4708,11 +4848,6 @@ class SemanticParser(BasicParser):
         if function_call_args is not None:
             assert is_inline
             found_func = False
-
-        current_class = expr.get_direct_user_nodes(lambda u: isinstance(u, ClassDef))
-        cls_name = current_class[0].name if current_class else None
-        if cls_name:
-            bound_class = self.scope.find(cls_name, 'classes', raise_if_missing = True)
 
         not_used = [d for d in decorators if d not in (*def_decorators.__all__, 'property')]
         if len(not_used) >= 1:
@@ -4798,7 +4933,9 @@ class SemanticParser(BasicParser):
             if is_interface:
                 name, _ = self.scope.get_new_incremented_symbol(interface_name, tmpl_idx)
 
-            scope = self.create_new_function_scope(name, decorators = decorators,
+            insertion_scope.python_names[name] = expr.name
+
+            scope = self.create_new_function_scope(expr.name, name, decorators = decorators,
                     used_symbols = expr.scope.local_used_symbols.copy(),
                     original_symbols = expr.scope.python_names.copy(),
                     symbolic_aliases = expr.scope.symbolic_aliases)
@@ -4819,7 +4956,7 @@ class SemanticParser(BasicParser):
                 if not is_compatible:
                     self.exit_function_scope()
                     # remove the new created scope and the function name
-                    self.scope.sons_scopes.pop(name)
+                    self.scope.sons_scopes.pop(expr.name)
                     if is_interface:
                         self.scope.remove_symbol(name)
                     continue
@@ -4852,7 +4989,7 @@ class SemanticParser(BasicParser):
             # to handle the case of a recursive function
             # TODO improve in the case of an interface
             recursive_func_obj = FunctionDef(name, arguments, [], results, scope = scope)
-            self.insert_function(recursive_func_obj)
+            self.insert_function(recursive_func_obj, insertion_scope)
 
             # Create a new list that store local variables for each FunctionDef to handle nested functions
             self._allocs.append(set())
@@ -4874,11 +5011,6 @@ class SemanticParser(BasicParser):
             for i in sub_funcs:
                 self._visit(i)
 
-            # Calling the Garbage collecting,
-            # it will add the necessary Deallocate nodes
-            # to the body of the function
-            body.insert2body(*self._garbage_collector(body))
-
             results = self._visit(results)
             if isinstance(results, EmptyNode):
                 results = FunctionDefResult(Nil())
@@ -4890,6 +5022,11 @@ class SemanticParser(BasicParser):
 
             self._check_pointer_targets(results_vars)
 
+            # Calling the Garbage collecting,
+            # it will add the necessary Deallocate nodes
+            # to the body of the function
+            body.insert2body(*self._garbage_collector(body))
+
             # Determine local and global variables
             global_vars = list(self.get_variables(self.scope.parent_scope))
             global_vars = [g for g in global_vars if body.is_user_of(g)]
@@ -4900,7 +5037,7 @@ class SemanticParser(BasicParser):
             imports   = list({imp:None for imp in imports}.keys())
 
             # remove the FunctionDef from the function scope
-            func_     = self.scope.functions.pop(name)
+            func_ = insertion_scope.functions.pop(name)
             is_recursive = False
             # check if the function is recursive if it was called on the same scope
             if func_.is_recursive and not is_inline:
@@ -5016,17 +5153,17 @@ class SemanticParser(BasicParser):
 
         if len(new_semantic_funcs) == 1 and not is_interface:
             new_semantic_funcs = new_semantic_funcs[0]
-            self.insert_function(new_semantic_funcs)
+            self.insert_function(new_semantic_funcs, insertion_scope)
         else:
             for f in new_semantic_funcs:
-                self.insert_function(f)
+                self.insert_function(f, insertion_scope)
 
             new_semantic_funcs = Interface(interface_name, new_semantic_funcs, syntactic_node=expr)
             if expr.python_ast:
                 new_semantic_funcs.set_current_ast(expr.python_ast)
             if cls_name:
                 bound_class.update_interface(expr, new_semantic_funcs)
-            self.insert_function(new_semantic_funcs)
+            self.insert_function(new_semantic_funcs, insertion_scope)
 
         return EmptyNode()
 
@@ -5086,7 +5223,7 @@ class SemanticParser(BasicParser):
 
         parent = self._find_superclasses(expr)
 
-        scope = self.create_new_class_scope(name, used_symbols=expr.scope.local_used_symbols,
+        cls_scope = self.create_new_class_scope(name, used_symbols=expr.scope.local_used_symbols,
                     original_symbols = expr.scope.python_names.copy())
 
         attribute_annotations = [self._visit(a) for a in expr.attributes]
@@ -5097,84 +5234,20 @@ class SemanticParser(BasicParser):
                         severity='error', symbol=a)
             else:
                 v = a[0]
-                scope.insert_variable(v)
+                cls_scope.insert_variable(v)
                 attributes.append(v)
+
+        self.exit_class_scope()
 
         docstring = self._visit(expr.docstring) if expr.docstring else expr.docstring
 
-        cls = ClassDef(name, attributes, [], superclasses=parent, scope=scope,
+        cls = ClassDef(name, attributes, [], superclasses=parent, scope=cls_scope,
                 docstring = docstring, class_type = dtype)
-        self.scope.parent_scope.insert_class(cls)
+        self.scope.insert_class(cls)
 
         methods = expr.methods
         for method in methods:
             cls.add_new_method(method)
-
-        syntactic_init_func = next((method for method in methods if method.name == '__init__'), None)
-        if syntactic_init_func is None:
-            argument = FunctionDefArgument(Variable(dtype, 'self', cls_base = cls), bound_argument = True)
-            self.scope.insert_symbol('__init__')
-            scope = self.create_new_function_scope('__init__')
-            scope.insert_variable(argument.var)
-            init_func = FunctionDef('__init__', [argument], (), cls_name=cls.name, scope=scope)
-            self.exit_function_scope()
-            self.insert_function(init_func)
-            cls.add_new_method(init_func)
-        else:
-            self._visit(syntactic_init_func)
-            init_func = self.scope.functions.pop('__init__')
-
-        if isinstance(init_func, Interface):
-            errors.report("Pyccel does not support interface constructor", symbol=init_func,
-                severity='fatal')
-
-        # create a new attribute to check allocation
-        deallocater_lhs = Variable(dtype, 'self', cls_base = cls, is_argument=True)
-        deallocater = DottedVariable(lhs = deallocater_lhs, name = self.scope.get_new_name('is_freed'),
-                                     class_type = PythonNativeBool(), is_private=True)
-        cls.add_new_attribute(deallocater)
-        deallocater_assign = Assign(deallocater, LiteralFalse())
-        init_func.body.insert2body(deallocater_assign, back=False)
-
-        syntactic_method = next((m for m in cls.methods if not m.is_semantic), None)
-        while syntactic_method:
-            self._visit(syntactic_method)
-            syntactic_method = next((m for m in cls.methods if not m.is_semantic), None)
-
-        syntactic_del_func = next((method for method in methods if method.name == '__del__'), None)
-        if syntactic_del_func is None:
-            argument = FunctionDefArgument(Variable(dtype, 'self', cls_base = cls), bound_argument = True)
-            self.scope.insert_symbol('__del__')
-            scope = self.create_new_function_scope('__del__')
-            scope.insert_variable(argument.var)
-            del_method = FunctionDef('__del__', [argument], [Pass()], scope=scope)
-            self.exit_function_scope()
-            self.insert_function(del_method)
-            cls.add_new_method(del_method)
-        else:
-            del_method = cls.get_method('__del__', expr)
-
-        # Add destructors to __del__ method
-        self._current_function = del_method.name
-        attribute = []
-        for attr in cls.attributes:
-            if not attr.on_stack:
-                attribute.append(attr)
-            elif isinstance(attr.class_type, CustomDataType) and not attr.is_alias:
-                attribute.append(attr)
-        if attribute:
-            # Create a new list that store local attributes
-            self._allocs.append(set())
-            self._pointer_targets.append({})
-            self._allocs[-1].update(attribute)
-            del_method.body.insert2body(*self._garbage_collector(del_method.body))
-            self._pointer_targets.pop()
-        condition = If(IfSection(PyccelNot(deallocater),
-                        [del_method.body]+[Assign(deallocater, LiteralTrue())]))
-        del_method.body = [condition]
-        self._current_function = None
-
-        self.exit_class_scope()
 
         return EmptyNode()
 
@@ -5502,7 +5575,7 @@ class SemanticParser(BasicParser):
         return Assert(test)
 
     def _visit_FunctionDefResult(self, expr):
-        f_name      = self._current_function
+        f_name      = self.current_function_name
         if isinstance(f_name, DottedName):
             f_name = f_name.name[-1]