+++ CHANGELOG.md
@@ -160,6 +160,11 @@ All notable changes to this project will be documented in this file.
 -   #2097 : Fix printing of an empty list.
 -   #2235 : Fix negative numbers in slice indices when translating to C.
 -   #2144 : Fix accidental imports due to modules making their contents public by default.
+-   #2312 : Fix rounding direction for negative integer elements in `np.linspace`.
+-   #2125 : Fix missing type check for argument of known type in a function with arguments whose type can be one of several types.
+-   #2042 : Fix missing index in loop unravelling for loop of length 1.
+-   #2093 : Fix scoping issue preventing class methods from sharing a name with locals in another class method.
+-   #1814 : Fix class method visitation order to correctly access the global scope from methods.
 -   #2261 : Add LLVM compilers.
 
 ### Changed
@@ -178,6 +183,7 @@ All notable changes to this project will be documented in this file.
 -   #2249 : Improve installation docs and recommend virtual environment.
 -   #2242 : Change format of compiler info files.
 -   #2302 : Print the deallocation in a 1 line if statement.
+-   #2125 : Add information about received data type to type errors when calling a function with the wrong type.
 -   \[INTERNALS\] `FunctionDef` is annotated when it is called, or at the end of the `CodeBlock` if it is never called.
 -   \[INTERNALS\] `InlinedFunctionDef` is only annotated if it is called.
 -   \[INTERNALS\] Build `utilities.metaclasses.ArgumentSingleton` on the fly to ensure correct docstrings.